requirements.txt
# Core dependencies
numpy>=1.20
matplotlib>=3.3
scipy>=1.7

# Configuration and validation
pydantic>=2.0
pyyaml>=6.0

# Progress bar
tqdm>=4.0

# Output formats
h5py>=3.0
imageio>=2.0
